ideal api
