+++ src/app/app.config.server.ts
@@ -10,28 +10,14 @@ import { mergeApplicationConfig, ApplicationConfig } from '@angular/core';
 import { provideServerRendering } from '@angular/platform-server';
 import { appConfig } from './app.config';
 
-import { TransferState } from './home/ServerService/TransferState' ;
-import { ExampleService } from './home/ServerService/example.service.browser';
-import { ExampleService as ExampleServiceServer } from './home/ServerService/example.service.server';
+import { ExampleService } from '@server/Example.service';
 
-import { ReflectiveInjector } from 'injection-js';
-
-export const transferState = new TransferState();
-
-// TODO: better angular di control
-// TODO: auto generate this
-export const injector = ReflectiveInjector.resolveAndCreate([
-  { provide: TransferState, useValue: transferState},
-  { provide: ExampleServiceServer, useClass: ExampleServiceServer },
-  { provide: ExampleService, useExisting: ExampleServiceServer },
-  { provide: 'ExampleService', useExisting: ExampleServiceServer }
-]);
 
 const serverConfig: ApplicationConfig = {
   providers: [
     provideServerRendering(),
     // TODO: auto generate this
-    { provide: ExampleService, useFactory: () => injector.get(ExampleService) }
+    { provide: ExampleService, useClass: ExampleService }
   ]
 };
