--- src/app/home/ServerService/example.service.browser.ts
@@ -1,10 +1,0 @@
-// TODO: auto generate this file
-import { Injectable } from "@angular/core";
-// very important for angular
-@Injectable({
-  providedIn: "root"
-})
-export class ExampleService {
-  _transferState: any;
-  async getTodo() {}
-}
